+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 - Consistent reset for settings.
 - Improved translation loading.
+- [WP_CLI] `m-archive` command have now a definition and all synopsis are up to date.
 
 ### Fixed
 - MailArchiver doesn't correctly honour previous error handler calls (thanks to [ajoah](https://github.com/ajoah)).